removed stupid stuff
git-svn-id: https://btanks.svn.sourceforge.net/svnroot/btanks/trunk@7253 5c0bf991-d514-0410-a65f-ce4832e571f3

+++ engine/src/alarm.cpp
@@ -39,7 +39,7 @@ const bool Alarm::tick(const float dt) {
 		if (_t < _period)
 			return false;
 			
-		long n = floorl(_t / _period);
+		int n = (int)floor(_t / _period);
 		_t -= _period * n;
 
 		while(_t > _period && _t > 0) //paranoid stuff
